requirements.txt
openai>=1.0.0
anthropic>=0.7.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
duckduckgo-search>=4.1.0
